packages/fell_loader/pyproject.toml
[project]
name = "fell-loader"
version = "0.1.0"
description = "Ingestion scripts for the fell-finder webapp"
authors = [{ name = "Ross Piper", email = "rp-webservices@outlook.com" }]
license = { text = "MIT License" }
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "bng-latlon>=1.1",
    "delta-spark>=4.0.0,<4.1",
    "diskcache>=5.6.3",
    "geopy>=2.4.1",
    "numpy<2.0",
    "polars>=1.17.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=18.1.0",
    "pyspark>=4.0.0,<4.1",
    "rasterio>=1.4.3",
    "setuptools<80.0",
    "tqdm>=4.67.1",
    "zstandard>=0.25.0",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[dependency-groups]
dev = ["tabulate>=0.9.0"]
